+++ back/src/forms/__tests__/helpers.ts
@@ -27,42 +27,6 @@ export const prepareRedis = async ({ emitterCompany, recipientCompany }) => {
     company: recipientCompany,
     companyTypes: ["WASTE_PROCESSOR"]
   });
-  // const emitterCompanyInfos = async () => [
-  //   {
-  //     id: emitterCompany.id,
-  //     siret: emitterCompany.siret,
-  //     securityCode: "1234",
-  //     companyTypes: ["PRODUCER"]
-  //   }
-  // ];
-
-  // await cachedGet(
-  //   emitterCompanyInfos,
-  //   COMPANY_INFOS_CACHE_KEY,
-  //   emitterCompany.siret,
-  //   {
-  //     parser: JSON,
-  //     options: { EX: 60 }
-  //   }
-  // );
-  // const recipientCompanyInfos = async () => [
-  //   {
-  //     id: recipientCompany.id,
-  //     siret: recipientCompany.siret,
-  //     securityCode: "1234",
-  //     companyTypes: ["WASTE_PROCESSOR"]
-  //   }
-  // ];
-
-  // await cachedGet(
-  //   recipientCompanyInfos,
-  //   COMPANY_INFOS_CACHE_KEY,
-  //   recipientCompany.siret,
-  //   {
-  //     parser: JSON,
-  //     options: { EX: 60 }
-  //   }
-  // );
 };
 
 export const prepareDB = async () => {