+++ back/src/forms/resolvers/mutations/markAsSealed.ts
@@ -3,7 +3,12 @@ import { MutationResolvers } from "../../../generated/graphql/types";
 import { getFormOrFormNotFound } from "../../database";
 import { expandFormFromDb } from "../../form-converter";
 import { checkCanMarkAsSealed } from "../../permissions";
-import { checkCanBeSealed, checkCompaniesType } from "../../validation";
+import {
+  beforeSignedByTransporterSchema,
+  checkCanBeSealed,
+  checkCompaniesType,
+  wasteDetailsSchema
+} from "../../validation";
 import transitionForm from "../../workflow/transitionForm";
 import { EventType } from "../../workflow/types";
 import { sendMail } from "../../../mailer/mailing";
@@ -26,6 +31,28 @@ const markAsSealedResolver: MutationResolvers["markAsSealed"] = async (
   await checkCanBeSealed(form);
 
   await checkCompaniesType(form);
+  let formUpdateInput = null;
+
+  if (
+    form.emitterIsForeignShip === true ||
+    form.emitterIsPrivateIndividual === true
+  ) {
+    // pre-validate when signature by producer will be by-passed at the end of this mutation
+    formUpdateInput = {
+      emittedAt: new Date(),
+      emittedBy: user.name,
+      emittedByEcoOrganisme: false,
+      // required for machine to authorize signature
+      emitterIsForeignShip: form.emitterIsForeignShip,
+      emitterIsPrivateIndividual: form.emitterIsPrivateIndividual
+    };
+    const futureForm: Form = {
+      ...form,
+      ...formUpdateInput
+    };
+    await wasteDetailsSchema.validate(futureForm);
+    await beforeSignedByTransporterSchema.validate(futureForm);
+  }
 
   const sealedForm = await transitionForm(user, form, {
     type: EventType.MarkAsSealed
@@ -40,14 +67,28 @@ const markAsSealedResolver: MutationResolvers["markAsSealed"] = async (
     await formRepository.updateAppendix2Forms(appendix2Forms);
   }
 
-  // send welcome email to emitter if it is not registered in TD
-  const emitterCompanyExists =
-    (await prisma.company.count({
-      where: { siret: form.emitterCompanySiret }
-    })) > 0;
+  if (form.emitterCompanySiret) {
+    // send welcome email to emitter if it is not registered in TD
+    const emitterCompanyExists =
+      (await prisma.company.count({
+        where: { siret: form.emitterCompanySiret }
+      })) > 0;
 
-  if (!emitterCompanyExists) {
-    await mailToNonExistentEmitter(sealedForm, formRepository);
+    if (!emitterCompanyExists) {
+      await mailToNonExistentEmitter(sealedForm, formRepository);
+    }
+  }
+
+  if (
+    formUpdateInput &&
+    (sealedForm.emitterIsForeignShip === true ||
+      sealedForm.emitterIsPrivateIndividual === true)
+  ) {
+    const updatedForm = await transitionForm(user, sealedForm, {
+      type: EventType.SignedByProducer,
+      formUpdateInput
+    });
+    return expandFormFromDb(updatedForm);
   }
 
   return expandFormFromDb(sealedForm);