+++ Changelog.md
@@ -20,6 +20,7 @@ TEMP_STORER_ACCEPTED -->|markAsResealed| RESEALED(RESEALED)
 TEMP_STORER_ACCEPTED -.->|markAsProcessed| PROCESSED(PROCESSED / AWAITING_GROUP / NO_TRACEABILITY)
 ```
 - Permettre aux forces de l'ordre d’accéder au pdf des bordereaux sur présentation d'un QR-code à durée de vie limitée [PR 1433](https://github.com/MTES-MCT/trackdechets/pull/1433)
+- En tant que particulier ou navire étranger je peux être producteur d'un BSDD [PR 1452](https://github.com/MTES-MCT/trackdechets/pull/1452)
 
 #### :bug: Corrections de bugs